Add media bottle support to graphing tool
Allow the user to define which media bottles are connected to which
vials. This way, media consumption per bottle can be calculated.

This commit adds a configuration file, a table with the bottle
consumption and an interactive interface with two modals:

- Edit bottle setup: user can write the list of media bottles,
their volume, and the vials they are connected to.

- See bottle setup: current setup is displayed, and user can notify
when they change a bottle. This info will be logged in the bottle.txt
configuration file.

The bottle configuration is stored in a txt file inside the experiment
folder and timestamps are added when bottles are changed. Whenever the
bottle/vial configuration is changed, a new file is created and former
setup is backed up.

Only the media consumed from the currently plugged bottle is calculated,
therefore allowing for a future notification system when a bottle is
running low.