New OD blank method
eVOLVER.py used to first calculate the blank OD value and then store it.
This commit adds data processing to store the raw blank and apply it
before calculating the OD value with the fit, reducing inaccuracies.

Summarized changes are:

- Take half of the function transform_data and place it in a new
function apply_OD_calibration (so each one does what they say), in order
to use the fit after performing the rest of calculations

- Store the raw value associated with 0 OD (raw 'blank' of the
calibration) when retrieving the active calibration

- Store the difference between the blank raw value of the calibration
and the blank raw value of the experiment

- Calculate the OD as a function of the raw value measured each time
plus the difference between calibration and experiment blank. That is:
f ( RawCal0 - RawExp0 + RawExpT ) = OD

- Add prompts so user can decide what blanking procedure to use