Optimize OD file loading (#26)
* Optimize OD file loading

Implement function similar to linux tail to read only the last OD values
used for dilution logic, instead of loading the whole file. This avoids
buildup of broadcast messages in long experiments with large OD log files.

* Revise requested changes

* Decrease BUFFER_SIZE

For current use cases, where a small number of lines is loaded,
a smaller BUFFER_SIZE makes file loading faster. For uses with larger
number of lines, this parameter can be passed as an argument.